# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeePlan API",
    "version": "1.0.0",
    "description": "API documentation for WeePlan Backend"
  },
  "servers": [
    {
      "url": "http://localhost:5013",
      "description": "Local server"
    },
    {
      "url": "http://3.226.133.13",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/category/": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Get all service categories",
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "summary": "Create new service category",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Venue Decoration"
                },
                "description": {
                  "type": "string",
                  "example": "Floral and lighting setup"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/category/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Get single category details",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "Category"
        ],
        "summary": "Update service category",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated Decoration Category"
                },
                "description": {
                  "type": "string",
                  "example": "Updated description"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "Delete service category",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/subcategory/": {
      "get": {
        "tags": [
          "Subcategory"
        ],
        "summary": "Get all subcategories",
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Subcategory"
        ],
        "summary": "Create new subcategory",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Floral Design"
                },
                "categoryId": {
                  "type": "string",
                  "example": "60d5ecb8b5c9c123456789ab"
                },
                "description": {
                  "type": "string",
                  "example": "Custom flower arrangements"
                }
              },
              "required": [
                "name",
                "categoryId"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/subcategory/by-category/{categoryId}": {
      "get": {
        "tags": [
          "Subcategory"
        ],
        "summary": "Get subcategories by parent category ID",
        "description": "",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/subcategory/{id}": {
      "get": {
        "tags": [
          "Subcategory"
        ],
        "summary": "Get single subcategory details",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "Subcategory"
        ],
        "summary": "Update subcategory",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated Subcategory Name"
                },
                "description": {
                  "type": "string",
                  "example": "Updated subcategory description"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Subcategory"
        ],
        "summary": "Delete subcategory",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/advisor/services": {
      "post": {
        "tags": [
          "Advisor"
        ],
        "summary": "Create advisor service (Admin)",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "1-on-1 Wedding Planning Consultation"
                },
                "description": {
                  "type": "string",
                  "example": "Expert advice on vendor selection and budgeting"
                },
                "price": {
                  "type": "number",
                  "example": 150
                }
              },
              "required": [
                "title",
                "description",
                "price"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/advisor/services/{id}": {
      "patch": {
        "tags": [
          "Advisor"
        ],
        "summary": "Update advisor service (Admin)",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Advisor"
        ],
        "summary": "Delete advisor service (Admin)",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}